backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
python-socketio==5.12.1
sendgrid==6.11.0
pytest==8.3.5
httpx==0.28.1
